requirements.txt
PyPDF2
python-dotenv==1.0.0
boto3 == 1.34.158
fastapi == 0.112.0
requests == 2.32.3
boto3 == 1.34.158
PyPDF2
uvicorn == 0.30.5
openai == 1.40.3
langchain == 0.2.12
pypdf == 4.3.1
langchain-community == 0.2.11
langchain_openai == 0.1.21
streamlit
reportlab
streamlit-option-menu
streamlit-authenticator
streamlit-extras
hydralit_components